+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Ops.hs
@@ -26,7 +26,7 @@ module Parsley.Internal.Backend.Machine.Ops (
     -- * Core Machine Operations
     dup, returnST,
     -- ** Abstracted Input Operations
-    sat, emitLengthCheck,
+    sat, emitLengthCheck, fetch,
     -- ** Register Operations
     newΣ, writeΣ, readΣ,
     -- ** Handler Operations
@@ -114,15 +114,23 @@ code to execute on failure, and a state @γ@, tries to read a character
 from the input within @γ@, executing the failure code if it does not
 exist or does not match.
 
-@since 1.0.0.0
+@since 1.5.0.0
 -}
-sat :: (?ops :: InputOps (Rep o))
-    => (Defunc Char -> Defunc Bool)        -- ^ Predicate to test the character with.
-    -> (Γ s o (Char : xs) n r a -> Code b) -- ^ Code to execute with updated state on success.
-    -> Code b                              -- ^ Code to execute on any failure.
-    -> Γ s o xs n r a                      -- ^ State @γ@ which contains the input.
+sat :: (Defunc Char -> Defunc Bool)                         -- ^ Predicate to test the character with.
+    -> ((Code Char -> Offset o -> aux -> Code b) -> Code b) -- ^ The source of the character
+    -> (Defunc Char -> Offset o -> aux -> Code b)           -- ^ Code to execute on success.
+    -> Code b                                               -- ^ Code to execute on failure.
     -> Code b
-sat p k bad γ@Γ{..} = next (offset input) $ \c offset' -> let v = FREEVAR c in _if (p v) (k (γ {operands = Op v operands, input = moveOne input offset'})) bad
+sat p src good bad = src $ \c input' aux -> let v = FREEVAR c in _if (p v) (good v input' aux) bad
+
+{-|
+Consumes the next character and adjusts the offset to match.
+
+@since 1.5.0.0
+-}
+fetch :: (?ops :: InputOps (Rep o))
+      => Offset o -> (Code Char -> Offset o -> Code b) -> Code b
+fetch input k = next (offset input) $ \c offset' -> k c (moveOne input offset')
 
 {-|
 Emits a length check for a number of characters \(n\) in the most efficient
@@ -241,6 +249,8 @@ not.
 
 @since 1.4.0.0
 -}
+--TODO: annoyingly, a `try` on its own binds a handler, even though it's footprint is negligable
+--      we should introduce a `noBinding` flag to the Always handler to mitigate this.
 bindAlwaysHandler :: forall s o xs n r a b. HandlerOps o
                   => Γ s o xs n r a                    -- ^ The state from which to capture the offset.
                   -> StaHandlerBuilder s o a           -- ^ The handler waiting to receive the captured offset and be bound.